pyproject.toml
[project]
name = "growlab"
version = "0.1.0"
description = "Living Light System — hybrid plant sculpture control software"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.19",
    "click>=8.1",
    "fastapi>=0.109",
    "jinja2>=3.1",
    "pyserial>=3.5",
    "python-multipart>=0.0.6",
    "smbus2>=0.4",
    "Pillow>=10.0",
    "uvicorn[standard]>=0.27",
]

[project.optional-dependencies]
pi = [
    "RPi.GPIO>=0.7",
    "picamera2>=0.3",
    "adafruit-circuitpython-seesaw>=1.14",
]
dev = [
    "httpx>=0.27",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
]

[project.scripts]
growlab = "pi.cli.main:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pi*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["pi"]
omit = ["pi/dashboard/templates/*"]
